fix(Tabs): 修复后缀布局问题
- 调整代码组织
- 优化「使用配置」demo 代码
