+++ components/tabs/style/index.less
@@ -73,12 +73,9 @@
             padding: 4px 16px;
         }
 
-        &-suffix {
-            flex: 1;
-        }
-
         &-wrapper {
             position: relative;
+            flex: 1;
             overflow: hidden;
             &::after,
             &::before {
