+++ docs/.vitepress/components/tabs/position.vue
@@ -1,29 +1,27 @@
 <template>
-    <FRadioGroup v-model="position">
-        <FRadio value="top">top</FRadio>
-        <FRadio value="bottom">bottom</FRadio>
-        <FRadio value="left">left</FRadio>
-        <FRadio value="right">right</FRadio>
-    </FRadioGroup>
+    <FRadioGroup
+        v-model="position"
+        :cancelable="false"
+        :options="
+            ['top', 'bottom', 'left', 'right'].map((i) => ({
+                label: i,
+                value: i,
+            }))
+        "
+    />
+    <FDivider />
+
     <FTabs :position="position" style="height: 300px">
         <FTabPane v-for="i in 20" :key="i" :name="'Tab ' + i" :value="i">
             <div class="tab-content">Tab {{ i }}</div>
         </FTabPane>
-        <template #prefix>Prefix</template>
-        <template #suffix>Suffix</template>
     </FTabs>
 </template>
 
-<script>
+<script setup>
 import { ref } from 'vue';
-export default {
-    setup() {
-        const position = ref('left');
-        return {
-            position,
-        };
-    },
-};
+
+const position = ref('left');
 </script>
 
 <style scoped>
